PythonCode/requirements.txt
dnspython==2.2.1
python-dotenv==1.0.0
web3==5.29.2
